Refactor mutation resolver functions into util
